+++ .idea/gradle.properties
@@ -1,2 +1,4 @@
 android.enableJetifier=true
+android.enableJetifier=true
+android.useAndroidX=true
 android.useAndroidX=true
