--- app/src/main/java/com/example/usha/login/LoginMainFragment.kt
@@ -1,35 +1,0 @@
-package com.example.usha.login
-
-import androidx.lifecycle.ViewModelProvider
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import com.example.usha.R
-
-class LoginMainFragment : Fragment() {
-
-//    companion object {
-//        fun newInstance() = LoginMainFragment()
-//    }
-
-    private lateinit var viewModel: LogingMainViewModel
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        return inflater.inflate(R.layout.fragment_login_main, container, false)
-    }
-
-    override fun onActivityCreated(savedInstanceState: Bundle?) {
-        super.onActivityCreated(savedInstanceState)
-        viewModel = ViewModelProvider(
-            this,
-            ViewModelProvider.NewInstanceFactory()
-        ).get(LogingMainViewModel::class.java)
-        // TODO: Use the ViewModel
-    }
-
-}
