+++ app/src/main/java/com/example/usha/logins/login/LoginFragment.kt
@@ -1,0 +1,27 @@
+package com.example.usha.logins.login
+
+import androidx.lifecycle.ViewModelProvider
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.example.usha.R
+
+class LoginFragment : Fragment() {
+
+    private lateinit var viewModel: LoginViewModel
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+
+        viewModel = ViewModelProvider(
+            this,
+            ViewModelProvider.NewInstanceFactory()
+        ).get(LoginViewModel::class.java)
+
+        return inflater.inflate(R.layout.fragment_login, container, false)
+    }
+}
