+++ app/src/main/res/layout/fragment_login_main.xml
@@ -1,35 +1,47 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical"
-    tools:context=".login.LoginMainFragment">
+    xmlns:tools="http://schemas.android.com/tools">
 
-    <androidx.appcompat.widget.Toolbar
-        android:id="@+id/toolbar4"
+    <data>
+        <variable
+            name="viewModel"
+            type="com.example.usha.logins.LogingMainViewModel" />
+    </data>
+
+    <LinearLayout
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:minHeight="?attr/actionBarSize"
-        android:theme="?attr/actionBarTheme">
+        android:layout_height="match_parent"
+        android:orientation="vertical"
+        tools:context=".logins.LoginMainFragment">
 
-        <ImageView
-            android:id="@+id/btnLoginMainX"
-            android:layout_width="wrap_content"
+        <androidx.appcompat.widget.Toolbar
+            android:id="@+id/toolbar4"
+            android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:src="@android:drawable/btn_dialog" />
-    </androidx.appcompat.widget.Toolbar>
+            android:minHeight="?attr/actionBarSize"
+            android:theme="?attr/actionBarTheme">
 
-    <Button
-        android:id="@+id/btnLoginMainLogin"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:text="로그인하기" />
+            <ImageView
+                android:id="@+id/btnLoginMainX"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:src="@android:drawable/btn_dialog"
+                android:onClick="@{(view)->viewModel.clickXbtn()}"/>
+        </androidx.appcompat.widget.Toolbar>
 
-    <Button
-        android:id="@+id/btnLoginMainregister"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:text="회원가입 하기" />
-</LinearLayout>
+        <Button
+            android:id="@+id/btnLoginMainLogin"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:text="로그인하기"
+            android:onClick="@{(view)->viewModel.clickLoginBtn()}"/>
+
+        <Button
+            android:id="@+id/btnLoginMainregister"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:text="회원가입 하기"
+            android:onClick="@{(view)->viewModel.clickRegisterBtn()}"/>
+    </LinearLayout>
+</layout>
